+++ daft/execution/native_executor.py
@@ -11,10 +11,10 @@ from daft.table import MicroPartition
 if TYPE_CHECKING:
     from daft.logical.builder import LogicalPlanBuilder
     from daft.runners.partitioning import (
+        LocalMaterializedResult,
         MaterializedResult,
         PartitionT,
     )
-    from daft.runners.pyrunner import PyMaterializedResult
 
 
 class NativeExecutor:
@@ -31,13 +31,13 @@ class NativeExecutor:
         psets: dict[str, list[MaterializedResult[PartitionT]]],
         daft_execution_config: PyDaftExecutionConfig,
         results_buffer_size: int | None,
-    ) -> Iterator[PyMaterializedResult]:
-        from daft.runners.pyrunner import PyMaterializedResult
+    ) -> Iterator[LocalMaterializedResult]:
+        from daft.runners.partitioning import LocalMaterializedResult
 
         psets_mp = {
             part_id: [part.micropartition()._micropartition for part in parts] for part_id, parts in psets.items()
         }
         return (
-            PyMaterializedResult(MicroPartition._from_pymicropartition(part))
+            LocalMaterializedResult(MicroPartition._from_pymicropartition(part))
             for part in self._executor.run(psets_mp, daft_execution_config, results_buffer_size)
         )
