+++ daft/runners/pyrunner.py
@@ -5,7 +5,6 @@ import logging
 import threading
 import uuid
 from concurrent import futures
-from dataclasses import dataclass
 from typing import TYPE_CHECKING, Iterator
 
 from daft.context import get_context
@@ -17,12 +16,12 @@ from daft.filesystem import glob_path_with_stats
 from daft.internal.gpu import cuda_device_count
 from daft.runners import runner_io
 from daft.runners.partitioning import (
+    LocalMaterializedResult,
+    LocalPartitionSet,
     MaterializedResult,
     PartialPartitionMetadata,
-    PartID,
     PartitionCacheEntry,
     PartitionMetadata,
-    PartitionSet,
 )
 from daft.runners.profiler import profiler
 from daft.runners.progress_bar import ProgressBar
@@ -43,69 +42,6 @@ ExecutionID = str
 
 # Unique ID for each task
 TaskID = str
-
-
-class LocalPartitionSet(PartitionSet[MicroPartition]):
-    _partitions: dict[PartID, MaterializedResult[MicroPartition]]
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._partitions = {}
-
-    def items(self) -> list[tuple[PartID, MaterializedResult[MicroPartition]]]:
-        return sorted(self._partitions.items())
-
-    def _get_merged_micropartition(self) -> MicroPartition:
-        ids_and_partitions = self.items()
-        assert ids_and_partitions[0][0] == 0
-        assert ids_and_partitions[-1][0] + 1 == len(ids_and_partitions)
-        return MicroPartition.concat([part.partition() for id, part in ids_and_partitions])
-
-    def _get_preview_micropartitions(self, num_rows: int) -> list[MicroPartition]:
-        ids_and_partitions = self.items()
-        preview_parts = []
-        for _, mat_result in ids_and_partitions:
-            part: MicroPartition = mat_result.partition()
-            part_len = len(part)
-            if part_len >= num_rows:  # if this part has enough rows, take what we need and break
-                preview_parts.append(part.slice(0, num_rows))
-                break
-            else:  # otherwise, take the whole part and keep going
-                num_rows -= part_len
-                preview_parts.append(part)
-        return preview_parts
-
-    def get_partition(self, idx: PartID) -> MaterializedResult[MicroPartition]:
-        return self._partitions[idx]
-
-    def set_partition(self, idx: PartID, part: MaterializedResult[MicroPartition]) -> None:
-        self._partitions[idx] = part
-
-    def set_partition_from_table(self, idx: PartID, part: MicroPartition) -> None:
-        self._partitions[idx] = PyMaterializedResult(part, PartitionMetadata.from_table(part))
-
-    def delete_partition(self, idx: PartID) -> None:
-        del self._partitions[idx]
-
-    def has_partition(self, idx: PartID) -> bool:
-        return idx in self._partitions
-
-    def __len__(self) -> int:
-        return sum(len(partition.partition()) for partition in self._partitions.values())
-
-    def size_bytes(self) -> int | None:
-        size_bytes_ = [partition.partition().size_bytes() for partition in self._partitions.values()]
-        size_bytes: list[int] = [size for size in size_bytes_ if size is not None]
-        if len(size_bytes) != len(size_bytes_):
-            return None
-        else:
-            return sum(size_bytes)
-
-    def num_partitions(self) -> int:
-        return len(self._partitions)
-
-    def wait(self) -> None:
-        pass
 
 
 class PyActorPool:
@@ -165,7 +101,9 @@ class PyActorPool:
         )
         new_part = partition.eval_expression_list(initialized_projection)
         return [
-            PyMaterializedResult(new_part, PartitionMetadata.from_table(new_part).merge_with_partial(partial_metadata))
+            LocalMaterializedResult(
+                new_part, PartitionMetadata.from_table(new_part).merge_with_partial(partial_metadata)
+            )
         ]
 
     def submit(
@@ -276,7 +214,7 @@ class PyRunner(Runner[MicroPartition], ActorPoolManager):
         self,
         builder: LogicalPlanBuilder,
         results_buffer_size: int | None = None,
-    ) -> Iterator[PyMaterializedResult]:
+    ) -> Iterator[LocalMaterializedResult]:
         # NOTE: Freeze and use this same execution config for the entire execution
         daft_execution_config = get_context().daft_execution_config
         execution_id = str(uuid.uuid4())
@@ -374,7 +312,7 @@ class PyRunner(Runner[MicroPartition], ActorPoolManager):
         self,
         execution_id: str,
         plan: physical_plan.MaterializedPhysicalPlan[MicroPartition],
-    ) -> Iterator[PyMaterializedResult]:
+    ) -> Iterator[LocalMaterializedResult]:
         local_futures_to_task: dict[futures.Future, PartitionTask] = {}
         pbar = ProgressBar(use_ray_tqdm=False)
 
@@ -393,7 +331,7 @@ class PyRunner(Runner[MicroPartition], ActorPoolManager):
                         break
 
                     elif isinstance(next_step, MaterializedResult):
-                        assert isinstance(next_step, PyMaterializedResult)
+                        assert isinstance(next_step, LocalMaterializedResult)
 
                         # A final result.
                         logger.debug("execution[%s] Yielding completed step", execution_id)
@@ -572,30 +510,7 @@ class PyRunner(Runner[MicroPartition], ActorPoolManager):
             partitions = instruction.run(partitions)
 
         results: list[MaterializedResult[MicroPartition]] = [
-            PyMaterializedResult(part, PartitionMetadata.from_table(part).merge_with_partial(partial))
+            LocalMaterializedResult(part, PartitionMetadata.from_table(part).merge_with_partial(partial))
             for part, partial in zip(partitions, final_metadata)
         ]
         return results
-
-
-@dataclass
-class PyMaterializedResult(MaterializedResult[MicroPartition]):
-    _partition: MicroPartition
-    _metadata: PartitionMetadata | None = None
-
-    def partition(self) -> MicroPartition:
-        return self._partition
-
-    def micropartition(self) -> MicroPartition:
-        return self._partition
-
-    def metadata(self) -> PartitionMetadata:
-        if self._metadata is None:
-            self._metadata = PartitionMetadata.from_table(self._partition)
-        return self._metadata
-
-    def cancel(self) -> None:
-        return None
-
-    def _noop(self, _: MicroPartition) -> None:
-        return None
